Support geo Uris as coordinate input format
Resolves https://github.com/MarcusWolschon/osmeditor4android/issues/2710
