+++ src/main/java/de/blau/android/GeoUrlActivity.java
@@ -2,7 +2,7 @@ package de.blau.android;
 
 import android.content.Intent;
 import android.net.Uri;
-import de.blau.android.util.GeoUrlData;
+import de.blau.android.util.GeoUriData;
 
 /**
  * Start vespucci with geo: URLs. see http://www.ietf.org/rfc/rfc5870.txt
@@ -13,7 +13,7 @@ public class GeoUrlActivity extends UrlActivity {
 
     @Override
     boolean setIntentExtras(Intent intent, Uri data) {
-        GeoUrlData geoUrlData = GeoUrlData.parse(data.getSchemeSpecificPart());
+        GeoUriData geoUrlData = GeoUriData.parse(data.getSchemeSpecificPart());
         if (geoUrlData != null) {
             intent.putExtra(GEODATA, geoUrlData);
             return true;
