+++ src/main/java/de/blau/android/ShareOnOpenStreetMap.java
@@ -4,7 +4,7 @@ import android.content.Intent;
 import android.net.Uri;
 import androidx.annotation.NonNull;
 import de.blau.android.contract.Urls;
-import de.blau.android.util.GeoUrlData;
+import de.blau.android.util.GeoUriData;
 
 /**
  * Take a geo intent and open the location on OSM
@@ -13,7 +13,7 @@ public class ShareOnOpenStreetMap extends IntentDataActivity {
 
     @Override
     protected void process(@NonNull Uri data) {
-        GeoUrlData geoUrlData = GeoUrlData.parse(data.getSchemeSpecificPart());
+        GeoUriData geoUrlData = GeoUriData.parse(data.getSchemeSpecificPart());
         if (geoUrlData != null) {
             double lat = geoUrlData.getLat();
             double lon = geoUrlData.getLon();
